% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faov.R
\name{faov}
\alias{faov}
\title{Export ANOVA summary results to Knitr/Sweave in APA format}
\usage{
faov(anova, rnb, apa = FALSE)
}
\arguments{
\item{anova}{A summary aov object.}

\item{rnb}{A digit indicating which row to format.}

\item{apa}{A boolean index to format results according to the APA 
guideline (remove 0 before the decimal point). Default to False.}
}
\value{
Return a string reporting an inline ANOVA results according
 to the APA guideline.
}
\description{
A function to format a given treatement (i.e. row) of an 
  ANOVA (functions aov, manova...) to a Markdown/LaTeX format according
  to the APA guidelines.
The output can be directly used in a Knitr/Sweave document.
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, University of de Montreal (Canada);
}
\seealso{
\code{\link{ftt}}, \code{\link{fmsd}}
}
\keyword{ANOVA,}
\keyword{APA,}
\keyword{Knitr/Sweave}

