% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fapa.R
\name{fapa}
\alias{fapa}
\title{Format a number according to the APA guidelines}
\usage{
fapa(nbr, p = FALSE)
}
\arguments{
\item{nbr}{A number to format.}

\item{p}{A boolean index to precise if the number to format is a p
value.}
}
\value{
Return a string reporting an inline number results according
 to the APA guideline.
}
\description{
A function to format a given number according to the APA
  guidelines, i.e. removing the 0 before the decimal point,
  and rounding the number to two digits.
If the number to format is a p value, then add  'p = ' or
  'p < .05'.
The output can be directly used in a Knitr/Sweave document.
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, University of de Montreal (Canada);
}
\seealso{
\code{\link{ftt}}, \code{\link{fmsd}}, \code{\link{faov}}
}
\keyword{APA,}
\keyword{Knitr/Sweave}
\keyword{p-value,}

