% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ftt}
\alias{ftt}
\title{Export t test results to APA format}
\usage{
ftt(ttest, d = NULL)
}
\arguments{
\item{ttest}{An t.test object.}

\item{d}{A number representing the associated d of Cohen.}
}
\description{
A function to format a t test of Student according to the
 APA guidelines. The output can be directly used in a
 Knitr/Sweave document.
The function can optionally export the related d of Cohen.
}
\examples{
# Generate fake data of two groups to be compared
cond1 = rnorm(10, mean=555, sd=45)
cond2 = rnorm(10, mean=639, sd=33)

# Compute the t test
ttest = t.test(cond1, cond2, paired=F)

# Export the t test result
ftt(ttest)
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, University of de Montreal (Canada);
}
\keyword{APA,}
\keyword{format,}
\keyword{knitr/Sweave}
\keyword{t}
\keyword{test,}

