% Generated by roxygen2 (4.0.0): do not edit by hand
\name{outliers}
\alias{outliers}
\title{Remove extreme with thresolds and values below and above the mean}
\usage{
outliers(data, fpass = NULL, target, sdv = 3)
}
\arguments{
\item{data}{A data frame in the long format (one row per record).}

\item{target}{A string indicating the column names of the data to filter.
Defaults to 'RT'.}

\item{fpass}{A vector length two with the minimal and maximal accepted
value to use for a first filtering of the reaction times.
To filter only the lowest or highest values, indicate NA as value.
For instance c(100, NA) will only remove RT < 100ms.
Defaults to NULL.}

\item{sdv}{A number indicating how many standard deviations should be
used to filter the data. Defaults to 3.}
}
\value{
Return a list with a data frame of filtered data and
 a data frame of number of data excluded and its relative percentage
 per condition.
}
\description{
A function to remove outliers of a given data frame. Extreme values
 can be excluded based on a minimal/maximal values and/or based on
 a given standard deviation value.
Return a data frame of filtered data as well as the number and
 relative percentage of data filtered.
}
\examples{
# Generate fake data with a subject number in the first colum,
#   a fake experimental condition in the second column and fake
#   reaction times in the third column
df = data.frame(Subj=1, Cond="Test", RT=rnorm(25, mean=550, sd=50))

# Adding extreme values
df[5,3]  <- df[5,3]+300
df[25,3] <- df[5,3]+500
df[19,3] <- df[19,3]-350
df[12,3] <- 55
df[15,3] <- 2340

# Filter with low and high thresolds and with 3 standard deviations
outliers(df, RT='RT', fpass=c(100,1000), sdv=3)

# Filter with only a low thresold with 2 standard deviations
outliers(df, target='RT', fpass=c(100, NA), sdv=2)
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, University of de Montreal (Canada);

Benoit A. Riou \email{riouba@gmail.com}, Lyon2 University (France)
}
\seealso{
\code{\link{filtRT}}
}
\keyword{filter,}
\keyword{outilers}
\keyword{reaction}
\keyword{times,}

