% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/faov.R
\name{faov}
\alias{faov}
\title{Export ezANOVA results to Knitr/Sweave in APA format}
\usage{
faov(anova, rnb, apa = FALSE)
}
\arguments{
\item{anova}{An ezANOVA object.}

\item{rnb}{A digit indicating which row to format.}

\item{apa}{A boolean index to format results according to the APA
guideline (remove 0 before the decimal point). Default to False.}
}
\value{
Return a string reporting an inline ANOVA results according
 to the APA guideline.
}
\description{
A function to format a given treatement (i.e. row) of an
  ezANOVA (package "ez") to a Markdown/LaTeX format according
  to the APA guidelines.
The output can be directly used in a Knitr/Sweave document.
The function tests if the Mauchly's test for sphericity is
 significant, and if so, report the corrected values.
}
\examples{
# Generate fake data with a subject number in the first colum, a
#  fake experimental condition in the second column and fake
#  reaction times in the third column
df = rbind(data.frame(Subj=1:10,  Cond="Control", RT=rnorm(10, mean=720, sd=59)),
           data.frame(Subj=11:20, Cond="Test", RT=rnorm(10, mean=680, sd=45)),
           data.frame(Subj=21:30, Cond="Test2", RT=rnorm(10, mean=650, sd=49)))
# Convert subject and independant variable to factors
df$Subj = as.factor(df$Subj)
df$Cond = as.factor(df$Cond)

# Compute the ANOVA
anova = ezANOVA(df, wid=.(Subj), between=.(Cond), dv=RT)

# Format the results
faov(anova, rnb=1)
}
\author{
Guillaume T. Vallet \email{gtvallet@gmail.com}, University of de Montreal (Canada);
}
\seealso{
\code{\link{ftt}}, \code{\link{fmsd}}
}
\keyword{ANOVA,}
\keyword{APA,}
\keyword{Knitr/Sweave}

